[fix] 게임 종료 후 메인으로 돌아갈 때 어치브먼트 토스트 꺼주기 #542
